requirements.txt
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=4.25.0
google-generativeai>=0.3.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
scikit-learn>=1.4.0
pandas>=2.2.0
Pillow>=10.2.0
flask>=3.0.0
twilio>=9.0.0